NaLang/corpus_manager/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
spacy>=3.7.0
nltk>=3.8.0
python-multipart>=0.0.9
pdfminer.six>=20221105
python-docx>=1.1.0
striprtf>=0.0.26
